removed build warning
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@1184 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ Subversion/subversion.cpp
@@ -599,7 +599,7 @@ void SubversionPlugin::DoMakeHTML(const wxArrayString &output, const wxString &b
 	}
 
 
-	for ( size_t i=0; i<book->GetPageCount(); i++) {
+	for ( size_t i=0; i<(size_t)book->GetPageCount(); i++) {
 		wxHtmlWindow *win = dynamic_cast<wxHtmlWindow *>(book->GetPage(i));
 		if (win && book->GetPageText(i) == wxT("SVN Status")) {
 			//we found a SVN status page, close it
